+++ IEntity.go
@@ -12,7 +12,7 @@ type IEntityStruct interface {
 	GetPKColumnName() string
 
 	//GetPkSequence 主键序列,因为需要兼容多种数据库的序列,所以使用map
-	//key是DBType,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
+	//key是Dialect,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
 	//如果key对应的value是 "",则代表是触发器触发的序列,兼容自增关键字,例如 ["oracle"]""
 	//GetPkSequence Primary key sequence, because it needs to be compatible with multiple database sequences, map is used
 	//The key is the DB Type, and the value is the value of the sequence,
@@ -35,7 +35,7 @@ type IEntityMap interface {
 	GetPKColumnName() string
 
 	//GetPkSequence 主键序列,因为需要兼容多种数据库的序列,所以使用map
-	//key是DBType,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
+	//key是Dialect,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
 	//如果key对应的value是 "",则代表是触发器触发的序列,兼容自增关键字,例如 ["oracle"]""
 	//GetPkSequence Primary key sequence, because it needs to be compatible with multiple database sequences, map is used
 	//The key is the DB Type, and the value is the value of the sequence,
@@ -78,7 +78,7 @@ func (entity *EntityStruct) GetPKColumnName() string {
 
 //var defaultPkSequence = make(map[string]string, 0)
 
-//GetPkSequence 主键序列,需要兼容多种数据库的序列,使用map,key是DBType,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
+//GetPkSequence 主键序列,需要兼容多种数据库的序列,使用map,key是Dialect,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
 //如果key对应的value是 "",则代表是触发器触发的序列,兼容自增关键字,例如 ["oracle"]""
 func (entity *EntityStruct) GetPkSequence() map[string]string {
 	return nil
@@ -92,7 +92,7 @@ type EntityMap struct {
 	tableName string
 	//主键列名
 	PkColumnName string
-	//主键序列,需要兼容多种数据库的序列,使用map,key是DBType,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
+	//主键序列,需要兼容多种数据库的序列,使用map,key是Dialect,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
 	PkSequence map[string]string
 	//数据库字段,不暴露外部
 	dbFieldMap map[string]interface{}
@@ -118,7 +118,7 @@ func (entity *EntityMap) GetPKColumnName() string {
 }
 
 //GetPkSequence 主键序列,因为需要兼容多种数据库的序列,所以使用map
-//key是DBType,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
+//key是Dialect,value是序列的值,例如oracle的TESTSEQ.NEXTVAL,如果有值,优先级最高
 //如果key对应的value是 "",则代表是触发器触发的序列,兼容自增关键字,例如 ["oracle"]""
 //GetPkSequence Primary key sequence, because it needs to be compatible with multiple database sequences, map is used
 //The key is the DB Type, and the value is the value of the sequence,